Merged in extract_config_from_global_var (pull request #6)
CLI: Refactor storing config as global variable

* CLI: Allow API server to be read from environment

* Merge branch 'master' into extract_config_from_global_var

* CLI: Refactor storing config as global variable

There were two options I could think of to remove the config from global scope:

1. Create a base class for Actions, defining config and gateway (so that
there is only a single place `ClientConfig.new` needs to be called

2. Call `ClientConfig.new` in Gateway (used by all Actions except
`Servers` as well as the Servers Action. The only risk here is if we
change the interface of `ClientConfig`, and we need to do something more
than `.new`ing it. I think the chance is low, so went for this option,
because I think it's simpler.